refact: refactor ios implementation
